requirements.txt
matplotlib==3.8.0
numpy==1.26.4
opencv-contrib-python-headless==4.11.0.86
opencv-python-headless==4.11.0.86
pillow==11.1.0
pycomm3==1.2.14
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
roboticstoolbox-python==1.1.1
rtb-data==1.0.1
scikit-image==0.25.2
scipy==1.15.2
setuptools==77.0.3
spatialgeometry==1.1.0
spatialmath-python==1.1.14
swift-sim==1.1.0
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
pyserial==3.5